Extract Redis client setup from authn into a separate module
I want to store "stale before" timestamps for users in Redis but don't
want to put those functions in the authn module.
