+++ src/authn/index.js
@@ -5,7 +5,6 @@ const assert = require("assert").strict;
 const debug = require("debug")("nextstrain:authn");
 const querystring = require("querystring");
 const session = require("express-session");
-const Redis = require("ioredis");
 const RedisStore = require("connect-redis")(session);
 const FileStore = require("session-file-store")(session);
 const passport = require("passport");
@@ -18,6 +17,7 @@ const BearerStrategy = require("./bearer");
 const {AuthnRefreshTokenInvalid} = require("../exceptions");
 const {fetch} = require("../fetch");
 const {copyCookie} = require("../middleware");
+const {REDIS} = require("../redis");
 const utils = require("../utils");
 
 const PRODUCTION = process.env.NODE_ENV === "production";
@@ -50,9 +50,6 @@ const SESSION_SECRET = PRODUCTION
   : "BAD SECRET FOR DEV ONLY";
 
 const SESSION_MAX_AGE = 30 * 24 * 60 * 60; // 30d in seconds
-
-/* Our heroku apps (nextstrain-dev & nextstrain-server) set the `REDIS_URL` env variable */
-const REDIS_URL = process.env.REDIS_URL;
 
 const COGNITO_USER_POOL_ID = "us-east-1_Cg5rcTged";
 
@@ -292,31 +289,11 @@ function setup(app) {
 
   // Setup session storage and passport.
   const sessionStore = () => {
-    if (REDIS_URL) {
-      const url = new URL(REDIS_URL);
-
-      // Heroku Redis' TLS socket is documented to be on the next port up.  The
-      // scheme for secure redis:// URLs is rediss://.
-      if (url.protocol === "redis:") {
-        utils.verbose(`Rewriting Redis URL <${scrubUrl(url)}> to use TLS`);
-        url.protocol = "rediss:";
-        url.port = Number(url.port) + 1;
-      }
-
-      utils.verbose(`Storing sessions in Redis at <${scrubUrl(url)}>`);
-
-      const client = new Redis(url.toString(), {
-        tls: {
-          // It is pretty frustrating that Heroku doesn't provide verifiable
-          // certs.  Although we're not using the Heroku Redis buildpack, the
-          // issue is documented here nicely
-          // <https://github.com/heroku/heroku-buildpack-redis/issues/15>.
-          rejectUnauthorized: false
-        }
-      });
+    if (REDIS) {
+      utils.verbose(`Storing sessions in Redis at <${REDIS.scrubbedUrl}>`);
 
       return new RedisStore({
-        client,
+        client: REDIS,
 
         // Using a key prefix reduces the chance that we conflict with
         // potential future usage of Redis in this codebase.  While this is
@@ -739,12 +716,6 @@ async function renewTokens(refreshToken) {
   };
 }
 
-
-function scrubUrl(url) {
-  const scrubbedUrl = new URL(url);
-  scrubbedUrl.password = "XXXXXX";
-  return scrubbedUrl;
-}
 
 module.exports = {
   setup
